src/voiceverse/tts.py
import torch
import torch.nn.functional as F
import numpy as np
from pathlib import Path
from typing import Union, Optional, Dict, Any
import yaml
import soundfile as sf
from transformers import AutoTokenizer, AutoModel

class StyleTTSWrapper:
    """StyleTTS2 wrapper with persona vector control"""
    
    def __init__(self, checkpoint_path: str, config_path: str, device: str = "cuda"):
        self.device = device
        self.checkpoint_path = Path(checkpoint_path)
        self.config_path = Path(config_path)
        
        # Load config
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)
        
        # Load model (placeholder - replace with actual StyleTTS2 loading)
        self.model = self._load_styletts2_model()
        
    def _load_styletts2_model(self):
        """Load StyleTTS2 model from checkpoint"""
        # Add actual StyleTTS2 model loading here
        # from StyleTTS2.models import build_model
        # model = build_model(self.config)
        # checkpoint = torch.load(self.checkpoint_path, map_location=self.device)
        # model.load_state_dict(checkpoint['model'])
        # return model.to(self.device).eval()
        return None  # placeholder
        
    def synthesize(self, text: str, persona_vector: torch.Tensor, 
                   style_controls: Dict[str, float] = None) -> np.ndarray:
        """
        Synthesize speech with persona vector and style controls
        
        Args:
            text: Input text (Telugu/English)
            persona_vector: Learned persona embedding
            style_controls: Dict with tempo, energy, warmth, formality controls
        """
        if style_controls is None:
            style_controls = {"tempo": 1.0, "energy": 1.0, "warmth": 0.5, "formality": 0.5}
        
        # Text preprocessing
        processed_text = self._preprocess_multilingual_text(text)
        
        # Generate mel-spectrogram with persona conditioning
        mel = self._generate_mel(processed_text, persona_vector, style_controls)
        
        return mel
    
    def _preprocess_multilingual_text(self, text: str) -> torch.Tensor:
        """Preprocess text for Telugu/English mixed input"""
        # Add language detection and phoneme conversion
        return torch.tensor([1, 2, 3])  # placeholder
    
    def _generate_mel(self, text_tensor: torch.Tensor, persona_vector: torch.Tensor, 
                      style_controls: Dict[str, float]) -> np.ndarray:
        """Generate mel-spectrogram with style control"""
        # Apply persona vector and style controls to model inference
        return np.random.randn(80, 100)  # placeholder mel

class ParlerTTSWrapper:
    """Parler-TTS wrapper for prompt-controlled generation"""
    
    def __init__(self, model_path: str):
        self.model_path = model_path
        # Load Parler-TTS model
        # self.tokenizer = AutoTokenizer.from_pretrained(model_path)
        # self.model = AutoModel.from_pretrained(model_path)
        
    def synthesize_with_prompt(self, text: str, description: str) -> np.ndarray:
        """
        Generate audio with natural language style description
        
        Args:
            text: Input text
            description: Style description like "warm, slow Telugu female voice"
        """
        # Implement Parler-TTS inference
        return np.random.randn(80, 100)  # placeholder
